pyproject.toml
[tool.poetry]
name = "SKIPP"
version = "0.1.0"
description = ""
authors = ["charbel-a-hC <charbel-a-h@outlook.com>"]
package-mode=false

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
matplotlib = "^3.7.2"
poethepoet = "^0.21.1"
seaborn = "^0.12.2"
tqdm = "^4.65.0"
networkx = "^3.1"
jupyter = "1.0.0"
wandb = "^0.15.6"
ipykernel = "^6.25.0"
PyVirtualDisplay = "^3.0"
pandoc = "^2.3"
Pillow = "8.*"
pyyaml = "^6.0.1"
einops = "^0.7.0"
numpy = "1.23"
yaml2pyclass = "^0.1.2"
scipy = "1.11"
evo = "^1.31.0"
scikit-learn = "^1.6.1"
scikit-image = "0.19.3"
opencv-python = "^4.11.0.86"
line-profiler = "^4.2.0"
huggingface-hub = "^0.29.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"] # Build tool dependency.
build-backend = "poetry.core.masonry.api"# Backend used to build the project.

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.12.0+cu113 torchvision==0.13.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html"

[tool.black]
line-length = 79

[tool.pycln]
all = true

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
